requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

asgiref>=3.8.1
Django>=5.1.7
sqlparse>=0.5.3
tzdata>=2025.1
google-generativeai>=0.8.3
python-docx>=1.1.2
torch>=2.7.1
torchvision>=0.22.1
Pillow>=10.0.0
sentence-transformers>=3.3.1
faiss-cpu>=1.9.0
pandas>=2.2.3
numpy>=2.1.3
scikit-learn>=1.6.0
gunicorn
whitenoise
